added-database-initializer

+++ src/SimpleStoreSample/SimpleStoreSample/Models/Product.cs
@@ -25,7 +25,7 @@ namespace SimpleStoreSample.Models
         public string ImagePath { get; set; }
 
         [Display(Name ="Price")]
-        public decimal? UnitPrice { get; set; }
+        public double? UnitPrice { get; set; }
 
         public int? CategoryID { get; set; }
